+++ windows/CMakeLists.txt
@@ -11,6 +11,7 @@ add_definitions(-DLIB_WEBRTC_API_DLL)
 add_library(${PLUGIN_NAME} SHARED
   "flutter_webrtc_plugin.cc"
   "src/flutter_webrtc.cc"
+  "src/flutter_peer_connection.cc"
   "${CMAKE_CURRENT_SOURCE_DIR}/rust/src/flutter_webrtc_native.cc"
 )
